Added arm64 support to SelectFolderViewController  Class

+++ Owncloud iOs Client/Tabs/FileTab/FileListDBOperations.h
@@ -41,7 +41,7 @@
  * Method that realice the refresh process
  *
  */
-+ (NSArray*)makeTheRefreshProcessWith:(NSMutableArray*)arrayFromServer inThisFolder:(int)idFolder;
++ (NSArray*)makeTheRefreshProcessWith:(NSMutableArray*)arrayFromServer inThisFolder:(NSInteger)idFolder;
 
 
 /*
